+++ src/main/scala/tigerjython/microbit/DeviceDetector.scala
@@ -1,0 +1,77 @@
+/*
+ * This file is part of the 'TigerJython' project.
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package tigerjython.microbit
+
+import tigerjython.utils.{OSPlatform, OSProcess}
+
+/**
+ * @author Tobias Kohn
+ */
+object DeviceDetector {
+
+  private var _calliopePath: String = _
+  private var _microBitPath: String = _
+
+  def getCalliopePath: String = _calliopePath
+
+  def getMicrobitPath: String = _microBitPath
+
+  def hasCalliopeReady: Boolean =
+    OSPlatform.system match {
+      case OSPlatform.WINDOWS =>
+        detectMicrobitWindows()
+        _calliopePath != null
+      case OSPlatform.LINUX | OSPlatform.MAC_OS =>
+        detectMicrobitPosix()
+        _calliopePath != null
+      case _ =>
+        _calliopePath = null
+        false
+    }
+
+  def hasMicrobitReady: Boolean =
+    OSPlatform.system match {
+      case OSPlatform.WINDOWS =>
+        detectMicrobitWindows()
+      case OSPlatform.LINUX | OSPlatform.MAC_OS =>
+        detectMicrobitPosix()
+      case _ =>
+        _microBitPath = null
+        false
+    }
+
+  /**
+   * Output of `mount`:
+   *   fs_spec on fs_file type fs_vfstype (fs_mntopts)
+   * We want the `fs_file` to check for `MICROBIT`.
+   * (Source: https://unix.stackexchange.com/questions/91960/can-anyone-explain-the-output-of-mount)
+   */
+  private def detectMicrobitPosix(): Boolean = {
+    _calliopePath = null
+    _microBitPath = null
+    val mountProcess = new OSProcess("mount")
+    mountProcess.exec()
+    for (line <- mountProcess.waitForOutput().split('\n')) {
+      val elements = line.split(' ')
+      if (elements.length >= 3 && elements(1) == "on") {
+        val fs_file = elements(2)
+        if (fs_file.endsWith("MICROBIT") && _microBitPath == null)
+          _microBitPath = fs_file
+        if (fs_file.endsWith("MINI") && _calliopePath == null)
+          _calliopePath = fs_file
+      }
+    }
+    _microBitPath != null
+  }
+
+  private def detectMicrobitWindows(): Boolean = {
+    _calliopePath = null
+    _microBitPath = null
+    false
+  }
+}